backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-jose[cryptography]
passlib[bcrypt]
aiohttp
beautifulsoup4
joblib
scikit-learn
pandas
dnspython
python-multipart
python-dotenv
